...Well...well...well
It's almost like AI models are incompetent liars, and not something anyone should use. Almost like putting AI in everything isn't a good idea because there are cool to use a toy but for anything practical they fucking suck because they can't think and are basically just plagiarizer 3000. 'Oh but, it's so good for making essaies', maybe it's because essays are easy to write

Anyway, hope this will be the last commit for the night (spoilers, it probably won't be because I haven't suffered enough)